move classes to another package
