+++ src/main/java/com/faforever/client/chat/ChatUserItemController.java
@@ -11,22 +11,22 @@ import com.faforever.client.game.GameTooltipController;
 import com.faforever.client.i18n.I18n;
 import com.faforever.client.player.PlayerService;
 import com.faforever.client.preferences.PreferencesService;
-import com.faforever.client.test.contextmenu.AddFoeMenuItem;
-import com.faforever.client.test.contextmenu.AddFriendMenuItem;
-import com.faforever.client.test.contextmenu.BroadcastMessageMenuItem;
-import com.faforever.client.test.contextmenu.ContextMenuBuilder;
-import com.faforever.client.test.contextmenu.CopyUsernameMenuItem;
-import com.faforever.client.test.contextmenu.InvitePlayerMenuItem;
-import com.faforever.client.test.contextmenu.JoinGameMenuItem;
-import com.faforever.client.test.contextmenu.KickGameMenuItem;
-import com.faforever.client.test.contextmenu.KickLobbyMenuItem;
-import com.faforever.client.test.contextmenu.RemoveFoeMenuItem;
-import com.faforever.client.test.contextmenu.RemoveFriendMenuItem;
-import com.faforever.client.test.contextmenu.ReportPlayerMenuItem;
-import com.faforever.client.test.contextmenu.SendPrivateMessageMenuItem;
-import com.faforever.client.test.contextmenu.ShowPlayerInfoMenuItem;
-import com.faforever.client.test.contextmenu.ViewReplaysMenuItem;
-import com.faforever.client.test.contextmenu.WatchGameMenuItem;
+import com.faforever.client.fx.contextmenu.AddFoeMenuItem;
+import com.faforever.client.fx.contextmenu.AddFriendMenuItem;
+import com.faforever.client.fx.contextmenu.BroadcastMessageMenuItem;
+import com.faforever.client.fx.contextmenu.ContextMenuBuilder;
+import com.faforever.client.fx.contextmenu.CopyUsernameMenuItem;
+import com.faforever.client.fx.contextmenu.InvitePlayerMenuItem;
+import com.faforever.client.fx.contextmenu.JoinGameMenuItem;
+import com.faforever.client.fx.contextmenu.KickGameMenuItem;
+import com.faforever.client.fx.contextmenu.KickLobbyMenuItem;
+import com.faforever.client.fx.contextmenu.RemoveFoeMenuItem;
+import com.faforever.client.fx.contextmenu.RemoveFriendMenuItem;
+import com.faforever.client.fx.contextmenu.ReportPlayerMenuItem;
+import com.faforever.client.fx.contextmenu.SendPrivateMessageMenuItem;
+import com.faforever.client.fx.contextmenu.ShowPlayerInfoMenuItem;
+import com.faforever.client.fx.contextmenu.ViewReplaysMenuItem;
+import com.faforever.client.fx.contextmenu.WatchGameMenuItem;
 import com.faforever.client.theme.UiService;
 import com.faforever.client.util.Assert;
 import com.google.common.eventbus.EventBus;
