+++ src/main/java/com/faforever/client/fx/contextmenu/AbstractCustomMenuItemController.java
@@ -1,4 +1,4 @@
-package com.faforever.client.test.contextmenu;
+package com.faforever.client.fx.contextmenu;
 
 import com.faforever.client.fx.Controller;
 import javafx.scene.control.CustomMenuItem;
