+++ src/main/java/com/faforever/client/fx/contextmenu/AddFoeMenuItem.java
@@ -1,4 +1,4 @@
-package com.faforever.client.test.contextmenu;
+package com.faforever.client.fx.contextmenu;
 
 import com.faforever.client.domain.PlayerBean;
 import com.faforever.client.i18n.I18n;
@@ -21,7 +21,7 @@ public class AddFoeMenuItem extends AbstractMenuItem<PlayerBean> {
   }
 
   @Override
-  protected boolean getVisible() {
+  protected boolean isItemVisible() {
     SocialStatus socialStatus = getObject().getSocialStatus();
     return socialStatus != FOE && socialStatus != SELF;
   }
