+++ src/main/java/com/faforever/client/fx/contextmenu/AddFriendMenuItem.java
@@ -1,4 +1,4 @@
-package com.faforever.client.test.contextmenu;
+package com.faforever.client.fx.contextmenu;
 
 import com.faforever.client.domain.PlayerBean;
 import com.faforever.client.i18n.I18n;
@@ -26,7 +26,7 @@ public class AddFriendMenuItem extends AbstractMenuItem<PlayerBean> {
   }
 
   @Override
-  protected boolean getVisible() {
+  protected boolean isItemVisible() {
     SocialStatus status = getObject().getSocialStatus();
     return status != FRIEND && status != SELF;
   }
