+++ src/main/java/com/faforever/client/leaderboard/SubDivisionTabController.java
@@ -8,10 +8,10 @@ import com.faforever.client.fx.JavaFxUtil;
 import com.faforever.client.fx.StringCell;
 import com.faforever.client.i18n.I18n;
 import com.faforever.client.notification.NotificationService;
-import com.faforever.client.test.contextmenu.ContextMenuBuilder;
-import com.faforever.client.test.contextmenu.CopyUsernameMenuItem;
-import com.faforever.client.test.contextmenu.ShowPlayerInfoMenuItem;
-import com.faforever.client.test.contextmenu.ViewReplaysMenuItem;
+import com.faforever.client.fx.contextmenu.ContextMenuBuilder;
+import com.faforever.client.fx.contextmenu.CopyUsernameMenuItem;
+import com.faforever.client.fx.contextmenu.ShowPlayerInfoMenuItem;
+import com.faforever.client.fx.contextmenu.ViewReplaysMenuItem;
 import com.faforever.client.theme.UiService;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.Tab;
