+++ src/main/java/com/faforever/client/teammatchmaking/PartyMemberItemController.java
@@ -11,12 +11,12 @@ import com.faforever.client.i18n.I18n;
 import com.faforever.client.leaderboard.LeaderboardService;
 import com.faforever.client.player.CountryFlagService;
 import com.faforever.client.player.PlayerService;
-import com.faforever.client.test.contextmenu.ContextMenuBuilder;
-import com.faforever.client.test.contextmenu.CopyUsernameMenuItem;
-import com.faforever.client.test.contextmenu.ReportPlayerMenuItem;
-import com.faforever.client.test.contextmenu.SendPrivateMessageMenuItem;
-import com.faforever.client.test.contextmenu.ShowPlayerInfoMenuItem;
-import com.faforever.client.test.contextmenu.ViewReplaysMenuItem;
+import com.faforever.client.fx.contextmenu.ContextMenuBuilder;
+import com.faforever.client.fx.contextmenu.CopyUsernameMenuItem;
+import com.faforever.client.fx.contextmenu.ReportPlayerMenuItem;
+import com.faforever.client.fx.contextmenu.SendPrivateMessageMenuItem;
+import com.faforever.client.fx.contextmenu.ShowPlayerInfoMenuItem;
+import com.faforever.client.fx.contextmenu.ViewReplaysMenuItem;
 import com.faforever.client.theme.UiService;
 import com.faforever.client.util.Assert;
 import com.faforever.commons.lobby.Faction;
