+++ src/main/resources/theme/avatar_picker_menu_item.fxml
@@ -1,6 +1,6 @@
 <?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.CustomMenuItem?>
-<CustomMenuItem fx:controller="com.faforever.client.test.contextmenu.AvatarPickerMenuItemController" fx:id="root"
+<CustomMenuItem fx:controller="com.faforever.client.fx.contextmenu.AvatarPickerMenuItemController" fx:id="root"
                 hideOnClick="false" mnemonicParsing="false" visible="false" xmlns:fx="http://javafx.com/fxml/1"
                 xmlns="http://javafx.com/javafx/8.0.141">
     <content>
