+++ src/main/resources/theme/color_picker_menu_item.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ColorPicker?>
 <?import javafx.scene.control.Button?>
-<CustomMenuItem fx:controller="com.faforever.client.test.contextmenu.ColorPickerMenuItemController" fx:id="root"
+<CustomMenuItem fx:controller="com.faforever.client.fx.contextmenu.ColorPickerMenuItemController" fx:id="root"
                 hideOnClick="false" mnemonicParsing="false" visible="false" xmlns:fx="http://javafx.com/fxml/1"
                 xmlns="http://javafx.com/javafx/8.0.141">
     <content>
